refact: Move Timer code from MainActivity to TimerFragment
- update MainActivity to launch TimerFragment
