+++ app/src/main/res/layout/activity_main.xml
@@ -1,81 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
 <android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".MainActivity">
 
-    <!-- Initial progress bar shown. Disappears when timer is started -->
-    <ProgressBar
-        android:id="@+id/pbar_static"
-        style="?android:attr/progressBarStyleHorizontal"
-        android:layout_width="360dp"
-        android:layout_height="360dp"
-        android:layout_marginStart="8dp"
-        android:layout_marginLeft="8dp"
-        android:layout_marginTop="80dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginRight="8dp"
-        android:indeterminate="false"
-        android:progressDrawable="@drawable/progress_bar_static"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <!-- Progress bar shown when timer starts and as timer updates -->
-    <ProgressBar
-        android:id="@+id/pbar_update"
-        style="?android:attr/progressBarStyleHorizontal"
-        android:layout_width="360dp"
-        android:layout_height="360dp"
-        android:layout_marginStart="8dp"
-        android:layout_marginLeft="8dp"
-        android:layout_marginTop="80dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginRight="8dp"
-        android:indeterminate="false"
-        android:progressDrawable="@drawable/progress_bar_update"
-        android:visibility="invisible"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:visibility="visible" />
-
-    <!-- Text for countdown timer -->
-    <TextView
-        android:id="@+id/tv_count_down_timer"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="00:00"
-        android:textColor="@android:color/black"
-        android:textSize="50sp"
-        app:layout_constraintBottom_toBottomOf="@+id/pbar_update"
-        app:layout_constraintEnd_toEndOf="@+id/pbar_update"
-        app:layout_constraintStart_toStartOf="@+id/pbar_update"
-        app:layout_constraintTop_toTopOf="@+id/pbar_update" />
-
-    <Button
-        android:id="@+id/btn_timer_start_pause"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="start"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.624" />
-
-    <Button
-        android:id="@+id/btn_timer_reset"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        android:text="reset"
-        android:visibility="invisible"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/btn_timer_start_pause"
-        tools:visibility="visible" />
+    <fragment
+        android:id="@+id/details_fragment"
+        android:name="com.example.namlu.focustimer.TimerFragment"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"/>
 
 </android.support.constraint.ConstraintLayout>
